refactor: Move the useModal hook to the ModalProvider file
