+++ src/components/SearchResult/useActions.jsx
@@ -12,7 +12,7 @@ import {
   viewInDrive
 } from 'src/components/Actions/Items'
 import { useFileSharing } from 'src/components/Contexts/FileSharingProvider'
-import { useModal } from 'src/components/Hooks/useModal'
+import { useModal } from 'src/components/Contexts/ModalProvider'
 import { useMultiSelection } from 'src/components/Hooks/useMultiSelection'
 
 import { isFile, isNote } from 'cozy-client/dist/models/file'
